java-junit/src/main/java/br/com/alura/tdd/service/BonusService.java
package br.com.alura.tdd.service;

import java.math.BigDecimal;
import java.math.RoundingMode;

import br.com.alura.tdd.modelo.Funcionario;

public class BonusService {

	public BigDecimal calcularBonus(Funcionario funcionario) {
		BigDecimal valor = funcionario.getSalario().multiply(new BigDecimal("0.1"));
		if (valor.compareTo(new BigDecimal("1000")) > 0) {
			throw new IllegalArgumentException("Salário muito alto para receber bônus");
		}
		return valor.setScale(2, RoundingMode.HALF_UP);
	}//garante que vai ter sempre 2 casas e arredonda pra cima

}
